+++ src/components/SearchFilter.vue
@@ -273,7 +273,7 @@ export default defineComponent({
     }, getDefaults());
   },
   computed: {
-    ...mapState(['searchResultsPerPage', 'maxEntriesPerPage', 'uiLanguage']),
+    ...mapState(['defaultCollectionSort', 'defaultItemSort', 'searchResultsPerPage', 'maxEntriesPerPage', 'uiLanguage']),
     ...mapGetters(['canSearchCollections', 'supportsConformance']),
     collectionSelectOptions() {
       let taggable = !this.hasAllCollections;
@@ -522,7 +522,10 @@ export default defineComponent({
           .catch(error => console.error(error))
       );
     }
-    Promise.all(promises).finally(() => this.loaded = true);
+    Promise.all(promises).finally(() => {
+      this.resetSort();
+      this.loaded = true;
+    });
   },
   methods: {
     resetSearchCollection() {
@@ -708,6 +711,7 @@ export default defineComponent({
     },
     async onReset() {
       Object.assign(this, getDefaults());
+      this.resetSort();
       this.$emit('input', this.query, true);
     },
     setLimit(limit) {
@@ -746,6 +750,20 @@ export default defineComponent({
       }
       else {
         return null;
+      }
+    },
+    resetSort() {
+      if (!this.canSort) {
+        return;
+      }
+      const defaults = this.type === 'Collections' ? this.defaultCollectionSort : this.defaultItemSort;
+      const sort = Utils.parseApiSortParameter(defaults);
+      if (sort.field) {
+        const sortOption = this.sortOptions.find(option => option.value === sort.field);
+        if (sortOption) {
+          this.sortTerm = sortOption;
+          this.sortOrder = sort.direction;
+        }
       }
     }
   }