+++ src/utils.js
@@ -170,7 +170,7 @@ export default class Utils {
     return pages;
   }
 
-  static addFiltersToLink(link, filters = {}, defaultLimit = null) {
+  static addFiltersToLink(link, filters = {}, defaultLimit = null, defaultSort = null) {
     let isEmpty = value => {
       return (value === null
       || (typeof value === 'number' && !Number.isFinite(value))
@@ -187,6 +187,9 @@ export default class Utils {
 
     if (typeof filters.limit !== 'number' && typeof defaultLimit === 'number') {
       filters.limit = defaultLimit;
+    }
+    if (typeof filters.sortby !== 'string' && typeof defaultSort === 'string') {
+      filters.sortby = defaultSort;
     }
 
     if (hasText(link.method) && link.method.toUpperCase() === 'POST') {
@@ -356,14 +359,18 @@ export default class Utils {
     return Utils.mergeDeep(target, ...sources);
   }
 
-  static convertHumanizedSortOrder(value) {
-    switch (value) {
-      case 'asc':
-        return 1;
-      case 'desc':
-        return -1;
-      default:
-        return 0;
+  static parseApiSortParameter(value) {
+    if (typeof value !== 'string') {
+      return { field: null, direction: 0 };
+    }
+    if (value.startsWith('-')) {
+      return { field: value.substring(1), direction: -1 };
+    }
+    else {
+      if (value.startsWith('+')) {
+        value = value.substring(1);
+      }
+      return { field: value, direction: 1 };
     }
   }